third_party/WebKit/LayoutTests/platform/win/external/wpt/referrer-policy/generic/subresource-test/area-navigate-expected.txt

This is a testharness.js-based test.
PASS Area is responding with HTTP headers 
Harness: the test ran to completion.
